requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.51.0
chromadb==0.5.15
sentence-transformers==3.1.1
langchain==0.3.1
langchain-community==0.3.1
langgraph==0.2.28
pypdf==4.3.1
pymupdf==1.24.10
aiofiles==24.1.0
httpx==0.27.2
pydantic==2.9.2
